docs: Add example usage of registerOptions
The usage/purpose of registerOptions is not immediately clear today. Understanding what "options" are, requires a deeper understanding of the RegisterProxyMixin.

This example is intended to help clear up the use of registerOptions. This leverages the same example types as registerOptionsForType, registers the "singleton: false" option for the connection after registering the factory.